refactor(mcp): extract module-level helpers to tools_helpers.py
